services_listar-pedidos/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.16.0
flask
firebase-admin

